Pin _batch_id as long at every F.lit emission site
Three remaining ``F.lit(batch_id)`` / ``F.lit(0)`` sites emitted
``_batch_id`` as int32 without an explicit ``.cast("long")``:

- ``_common.py::_add_cdc_metadata`` (single-batch insert/update/delete
  per-batch metadata attach)
- ``api.py`` empty-batch fallback for Faker-containing tables
- ``single_batch.py::generate_initial_snapshot`` (batch 0 snapshot)

Commits 8e96ef5 and 46b7849 established "``_batch_id`` is long
everywhere else in the engine" -- fused.py, bulk.py, seed
derivation, CDC output schema.  These three sites worked today
because Spark's int -> long promotion kicks in on the union across
tables whose ``_batch_id`` is already long, but the drift from the
documented invariant was flagged in review.  Pin long explicitly
at each emission so the schema unifies without relying on
promotion; future callers reading ``_batch_id`` bytes (e.g. Delta
schema evolution) see a consistent long everywhere.

+++ dbldatagen/core/engine/cdc/_common.py
@@ -178,9 +178,15 @@ def _add_cdc_metadata(
     """
     if df is None:
         return None
+    # ``_batch_id`` is long everywhere else in the engine (fused.py,
+    # bulk.py, CDC output schema, seed derivation).  ``F.lit(int)``
+    # without the explicit cast widens to int32; downstream unions
+    # across tables with long _batch_id widen to long via promotion
+    # today, but the inconsistency drifts from the "long end-to-end"
+    # invariant committed in 8e96ef5 / 46b7849.  Pin long here.
     return (
         df.withColumn("_op", F.lit(op))
-        .withColumn("_batch_id", F.lit(batch_id))
+        .withColumn("_batch_id", F.lit(batch_id).cast("long"))
         .withColumn("_ts", F.lit(batch_ts_epoch).cast("long").cast("timestamp"))
     )
